# docs/openapi/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"LegacyFwcloudResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"created_by": {
						"type": "number",
						"format": "double"
					},
					"updated_by": {
						"type": "number",
						"format": "double"
					},
					"locked_at": {
						"type": "string",
						"nullable": true
					},
					"locked_by": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"locked": {
						"type": "number",
						"format": "double"
					},
					"image": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"created_at",
					"updated_at",
					"created_by",
					"updated_by",
					"locked_at",
					"locked_by",
					"locked",
					"image",
					"comment"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFwcloudErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFirewallInsertResponse": {
				"properties": {
					"insertId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"insertId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFirewallErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFirewallStoreRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"save_user_pass": {
						"type": "number",
						"format": "double"
					},
					"install_port": {
						"type": "number",
						"format": "double"
					},
					"fwmaster": {
						"type": "number",
						"format": "double"
					},
					"options": {
						"type": "number",
						"format": "double"
					},
					"node_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud",
					"name",
					"save_user_pass",
					"install_port",
					"fwmaster",
					"options",
					"node_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFirewallUpdateRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					},
					"firewall": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"save_user_pass": {
						"type": "number",
						"format": "double"
					},
					"install_port": {
						"type": "number",
						"format": "double"
					},
					"fwmaster": {
						"type": "number",
						"format": "double"
					},
					"options": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud",
					"firewall",
					"name",
					"comment",
					"save_user_pass",
					"install_port",
					"fwmaster",
					"options"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyFirewallGetRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					},
					"firewall": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud",
					"firewall"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LegacyFirewallCloudRequest": {
				"properties": {
					"fwcloud": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fwcloud"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LegacyClusterErrorResponse": {
				"properties": {
					"fwcErr": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "FWCloud API",
		"version": "2.4.1",
		"description": "FWCloud RESTful API.",
		"contact": {}
	},
	"paths": {
...
		"/fwcloud/all/get": {
			"get": {
				"operationId": "GetAllowed",
				"responses": {
					"200": {
						"description": "Get allowed fwclouds",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFwcloudResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 4,
												"name": "FWCloud-02",
												"created_at": "2019-05-14T11:37:19.000Z",
												"updated_at": "2019-05-14T11:37:19.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": null,
												"locked_by": null,
												"locked": 0,
												"image": "",
												"comment": ""
											},
											{
												"id": 5,
												"name": "FWCloud-03",
												"created_at": "2019-05-14T11:37:24.000Z",
												"updated_at": "2019-05-14T11:57:06.000Z",
												"created_by": 0,
												"updated_by": 0,
												"locked_at": "2019-05-14T11:57:06.000Z",
												"locked_by": 1,
												"locked": 1,
												"image": "",
												"comment": ""
											}
										]
									}
								}
							}
						}
					},
					"204": {
						"description": "No allowed fwclouds"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFwcloudErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7000,
											"msg": "FWCloud access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get fwcloud data for all the fwclouds to which the logged user has access.",
				"summary": "Get allowed fwclouds.",
				"tags": [
					"fwcloud"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/firewall": {
			"post": {
				"operationId": "Store",
				"responses": {
					"200": {
						"description": "New firewall",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallInsertResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"insertId": 1
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 1002,
											"msg": "Not found"
										}
									}
								}
							}
						}
					},
					"default": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Tree node access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new firewall.",
				"summary": "New firewall.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallStoreRequest"
							},
							"example": {
								"fwcloud": 1,
								"name": "Firewall-01",
								"save_user_pass": 0,
								"install_port": 22,
								"fwmaster": 0,
								"options": 0,
								"node_id": 1
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"204": {
						"description": "Update firewall"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update firewall information.",
				"summary": "Update firewall.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallUpdateRequest"
							},
							"example": {
								"fwcloud": 1,
								"firewall": 5,
								"name": "Firewall-UPDATED",
								"comment": "Comment for the updated firewall.",
								"save_user_pass": 0,
								"install_port": 22,
								"fwmaster": 0,
								"options": 3
							}
						}
					}
				}
			}
		},
		"/firewall/get": {
			"put": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Get firewall data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 5,
											"cluster": null,
											"fwcloud": 1,
											"name": "Firewall-05",
											"comment": null,
											"created_at": "2019-05-15T10:34:46.000Z",
											"updated_at": "2019-05-15T10:34:47.000Z",
											"compiled_at": null,
											"installed_at": null,
											"by_user": 1,
											"status": 3,
											"install_user": "",
											"install_pass": "",
											"save_user_pass": 0,
											"install_interface": null,
											"install_ipobj": null,
											"fwmaster": 0,
											"install_port": 22,
											"options": 0,
											"interface_name": null,
											"ip_name": null,
											"ip": null,
											"id_fwmaster": null
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewall data.",
				"summary": "Get firewall data.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallGetRequest"
							},
							"example": {
								"fwcloud": 1,
								"firewall": 5
							}
						}
					}
				}
			}
		},
		"/firewall/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get firewalls in cloud",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFirewallResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-01",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											},
											{
												"id": 2,
												"cluster": null,
												"fwcloud": 1,
												"name": "Firewall-02",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewalls data by fwcloud.",
				"summary": "Get firewalls in cloud.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallCloudRequest"
							},
							"example": {
								"fwcloud": 1
							}
						}
					}
				}
			}
		},
		"/firewall/cluster/get": {
			"put": {
				"operationId": "GetByCluster",
				"responses": {
					"200": {
						"description": "Get firewalls in cluster",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyFirewallResponse"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"cluster": 2,
												"fwcloud": 1,
												"name": "Firewall-01",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 1,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											},
											{
												"id": 2,
												"cluster": 2,
												"fwcloud": 1,
												"name": "Firewall-02",
												"comment": null,
												"created_at": "2019-05-15T10:34:46.000Z",
												"updated_at": "2019-05-15T10:34:47.000Z",
												"compiled_at": null,
												"installed_at": null,
												"by_user": 1,
												"status": 3,
												"install_user": "",
												"install_pass": "",
												"save_user_pass": 0,
												"install_interface": null,
												"install_ipobj": null,
												"fwmaster": 0,
												"install_port": 22,
												"options": 0,
												"interface_name": null,
												"ip_name": null,
												"ip": null,
												"id_fwmaster": null
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get firewalls data by cluster.",
				"summary": "Get firewalls in cluster.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallClusterRequest"
							},
							"example": {
								"fwcloud": 1,
								"cluster": 2
							}
						}
					}
				}
			}
		},
		"/firewall/clone": {
			"put": {
				"operationId": "Clone",
				"responses": {
					"200": {
						"description": "Clone firewall",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallInsertResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"insertId": 7
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create a new firewall cloning the one indicated in the request's parameters.",
				"summary": "Clone firewall.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallCloneRequest"
							},
							"example": {
								"fwcloud": 1,
								"firewall": 5,
								"name": "Firewall-CLONED",
								"comment": "Comment for the cloned firewall.",
								"node_id": 1
							}
						}
					}
				}
			}
		},
		"/firewall/del": {
			"put": {
				"operationId": "Destroy",
				"responses": {
					"204": {
						"description": "Delete firewall"
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyFirewallErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7001,
											"msg": "Firewall access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete a firewall.",
				"summary": "Delete firewall.",
				"tags": [
					"firewall"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyFirewallGetRequest"
							},
							"example": {
								"fwcloud": 1,
								"firewall": 7
							}
						}
					}
				}
			}
		},
...
		"/cluster/get": {
			"put": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Get cluster data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 2,
											"fwcloud": 2,
											"name": "Cluster-02",
											"comment": null,
											"created_at": "2019-05-17T11:47:00.000Z",
											"updated_at": "2019-05-17T11:47:00.000Z",
											"created_by": 0,
											"updated_by": 0,
											"nodes": [
												{
													"id": 10,
													"cluster": 2,
													"name": "node1",
													"comment": null,
													"status": 3,
													"install_user": "",
													"install_pass": "",
													"save_user_pass": 1,
													"install_interface": null,
													"install_ipobj": null,
													"fwmaster": 1,
													"install_port": 22,
													"interface_name": null,
													"ip_name": null,
													"ip": null,
													"options": 3
												}
											]
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Cluster access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get cluster data.",
				"summary": "Get cluster data.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetRequest"
							},
							"example": {
								"fwcloud": 2,
								"cluster": 5
							}
						}
					}
				}
			}
		},
		"/cluster/cloud/get": {
			"put": {
				"operationId": "GetByCloud",
				"responses": {
					"200": {
						"description": "Get cloud clusters",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LegacyClusterSummary"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": 1,
												"fwcloud": 2,
												"name": "Cluster-01",
												"comment": null,
												"created_at": "2019-05-17T11:46:57.000Z",
												"updated_at": "2019-05-17T11:46:57.000Z",
												"created_by": 0,
												"updated_by": 0
											},
											{
												"id": 2,
												"fwcloud": 2,
												"name": "Cluster-02",
												"comment": null,
												"created_at": "2019-05-17T11:47:00.000Z",
												"updated_at": "2019-05-17T11:47:00.000Z",
												"created_by": 0,
												"updated_by": 0
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Legacy endpoint error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LegacyClusterErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"fwcErr": 7002,
											"msg": "Cluster access not allowed"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all the cluster data for the indicated fwcloud.",
				"summary": "Get cloud clusters.",
				"tags": [
					"cluster"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LegacyClusterGetByCloudRequest"
							},
							"example": {
								"fwcloud": 2
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}